Remove usercontroller dependency from internal
