+++ packages/server/src/sdk/users/utils.ts
@@ -64,6 +64,25 @@ export async function rawUserMetadata(db?: Database) {
   ).rows.map(row => row.doc)
 }
 
+export async function fetchMetadata() {
+  const global = await getGlobalUsers()
+  const metadata = await rawUserMetadata()
+  const users = []
+  for (let user of global) {
+    // find the metadata that matches up to the global ID
+    const info = metadata.find(meta => meta._id.includes(user._id))
+    // remove these props, not for the correct DB
+    users.push({
+      ...user,
+      ...info,
+      tableId: InternalTables.USER_METADATA,
+      // make sure the ID is always a local ID, not a global one
+      _id: generateUserMetadataID(user._id),
+    })
+  }
+  return users
+}
+
 export async function syncGlobalUsers() {
   // sync user metadata
   const dbs = [context.getDevAppDB(), context.getProdAppDB()]
